ADM-402 move pipelineTool store to config store
