+++ frontend/__tests__/src/utils/setupStoreUtil.tsx
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit'
 import { stepperSlice } from '@src/context/stepper/StepperSlice'
 import { configSlice } from '@src/context/config/configSlice'
 import { boardSlice } from '@src/context/config/board/boardSlice'
-import { pipelineToolSlice } from '@src/context/pipelineTool/pipelineToolSlice'
+import { pipelineToolSlice } from '@src/context/config/pipelineTool/pipelineToolSlice'
 import { jiraVerifyResponseSlice } from '@src/context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
 import { sourceControlSlice } from '@src/context/sourceControl/sourceControlSlice'
 import { metricsSlice } from '@src/context/Metrics/metricsSlice'
