+++ frontend/src/store.ts
@@ -2,7 +2,7 @@ import { configureStore } from '@reduxjs/toolkit'
 import stepperReducer from './context/stepper/StepperSlice'
 import configReducer from './context/config/configSlice'
 import jiraVerifyResponseReducer from './context/config/board/jiraVerifyResponse/jiraVerifyResponseSlice'
-import pipelineToolResponseReducer from './context/pipelineTool/pipelineToolVerifyResponse/pipelineToolVerifyResponseSlice'
+import pipelineToolResponseReducer from './context/config/pipelineTool/pipelineToolVerifyResponse/pipelineToolVerifyResponseSlice'
 import sourceControlReducer from './context/sourceControl/sourceControlSlice'
 import sourceControlVerifyResponseReducer from './context/sourceControl/sourceControlVerifyResponse/sourceControlVerifyResponseSlice'
 import saveMetricsSettingReducer from './context/Metrics/metricsSlice'
